Add foreman_ygg_compat package
This package provides a compatibility layer to have the remote execution
pull provider be able to run with both yggdrasil 0.2.z and 0.4.z.

It creates a new dummy yggdrasild.service as the mangement piece for
the new yggdrasil services. Because of the BindsTo=, the new services
are started when this one is, because of the PartOf= overrides in
the other two services, the other two services are restarted when
yggdrasild.service is.

It also places a symlink so that the yggd utility can be invoked as
yggdrasil as was previously possible with the version of yggdrasil we
shipped.